requirements.txt
# InstaHotel - Dependencies

# AI
anthropic>=0.40

# Database
supabase>=2.3.0

# Google Drive API
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Image processing
Pillow>=10.0.0

# Video processing
opencv-python>=4.8.0

# Data models
pydantic>=2.0.0

# Environment
python-dotenv>=1.0.0
